nlp-service/requirements.txt
fastapi
uvicorn
spacy
mysql-connector-python
pydantic
